+++ src/main/java/com/example/petmanagment/Pet.java
@@ -1,9 +1,7 @@
 package com.example.petmanagment;
 
-
 import lombok.Getter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -12,9 +10,9 @@ public class Pet {
     private String animalSpecies;
     private String name;
     private LocalDate birthday;
-    private BigDecimal weight;
+    private Double weight;
 
-    public Pet(Integer id, String animalSpecies, String name, LocalDate birthday, BigDecimal weight) {
+    public Pet(Integer id, String animalSpecies, String name, LocalDate birthday, Double weight) {
         this.id = id;
         this.animalSpecies = animalSpecies;
         this.name = name;
@@ -22,7 +20,7 @@ public class Pet {
         this.weight = weight;
     }
 
-    public Pet(String animalSpecies, String name, LocalDate birthday, BigDecimal weight) {
+    public Pet(String animalSpecies, String name, LocalDate birthday, Double weight) {
         this.id = null;
         this.animalSpecies = animalSpecies;
         this.name = name;
